Small fix in socket_functions.c. Commencing irc-bot.
I realized that if there were to be an error in the call to getaddrinfo
the sockfd would be 1, aka stdout. This would be a serious issue and
would cause unknown issues.

I've started work on the irc bot MrOil requested. I also realize dial
(src/irc/utils.c) and sockets_connect (src/socket_functions.c) are the
same except for the call. I plan to modify the both of them to handle
both calls from both programs. I'll probably create a shared/ folder to
hold any source that both contain.